# api/openapi.yaml
openapi: 3.0.3
info:
  title: Preview Catalog API
  version: 1.0.0
  description: 'REST API for the Preview Platform demo application.

    Exposes a product catalogue with categories, reviews, orders and statistics.

    Used by Microcks for contract validation in every preview environment.

    '
  contact:
    name: ihsenalaya
    url: https://github.com/ihsenalaya/idp-testing
servers:
- url: http://svc-backend:8080
  description: In-cluster backend (preview namespace)
- url: http://localhost:8080
  description: Local development
paths:
...
  /api/categories:
    get:
      operationId: listCategories
      summary: List all categories with product count
      tags:
      - Categories
      responses:
        '200':
          description: Array of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
              examples:
                Example_1:
                  value:
                  - id: 1
                    name: Electronics
                    slug: electronics
                    product_count: 5
                  - id: 2
                    name: Clothing
                    slug: clothing
                    product_count: 3
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createCategory
      summary: Create a new category
      tags:
      - Categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
            examples:
              Example_1:
                value:
                  name: Electronics
                  slug: electronics
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryCreated'
              examples:
                Example_1:
                  value:
                    id: 1
                    name: Electronics
                    slug: electronics
        '400':
          description: Validation error — name and slug are required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: name and slug are required
        '500':
          $ref: '#/components/responses/InternalError'
  /api/products:
    get:
      operationId: listProducts
      summary: List products (last 50, with average rating)
      tags:
      - Products
      responses:
        '200':
          description: Array of products with ratings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
              examples:
                Example_1:
                  value:
                  - id: 1
                    name: Wireless Headphones
                    description: Premium noise-cancelling headphones
                    price: 199.99
                    stock: 12
                    discount_pct: 10.0
                    created_at: '2026-05-01T10:00:00+00:00'
                    category_name: Electronics
                    avg_rating: 4.5
                    review_count: 6
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createProduct
      summary: Create a new product
      tags:
      - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
            examples:
              Example_1:
                value:
                  name: Wireless Headphones
                  description: Premium noise-cancelling headphones
                  category_id: 1
                  price: 199.99
                  stock: 12
                  discount_pct: 10.0
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCreated'
              examples:
                Example_1:
                  value:
                    id: 1
                    name: Wireless Headphones
                    price: 199.99
                    stock: 12
                    discount_pct: 10.0
                    created_at: '2026-05-01T10:00:00+00:00'
        '400':
          description: Validation error — name and price are required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: name and price are required
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/products/{id}:
    get:
      operationId: getProduct
      summary: Get a product by ID with its reviews
      tags:
      - Products
      parameters:
      - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product with reviews
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
              examples:
                Example_1:
                  value:
                    id: 1
                    name: Wireless Headphones
                    description: Premium noise-cancelling headphones
                    price: 199.99
                    stock: 12
                    discount_pct: 10.0
                    created_at: '2026-05-01T10:00:00+00:00'
                    category_name: Electronics
                    reviews:
                    - id: 1
                      author: alice
                      rating: 5
                      comment: Excellent product
                      created_at: '2026-05-02T08:00:00+00:00'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deleteProduct
      summary: Delete a product
      tags:
      - Products
      parameters:
      - $ref: '#/components/parameters/ProductId'
      responses:
        '204':
          description: Product deleted (no body)
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/stats:
    get:
      operationId: getStats
      summary: Aggregated platform statistics
      tags:
      - Stats
      responses:
        '200':
          description: Platform statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stats'
              examples:
                Example_1:
                  value:
                    total_products: 10
                    total_categories: 3
                    total_reviews: 20
                    total_orders: 5
                    out_of_stock: 1
                    low_stock: 2
                    avg_rating: 4.2
                    categories:
                    - Electronics
                    - Clothing
                    - Books
        '500':
          $ref: '#/components/responses/InternalError'
components:
  parameters:
    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Product identifier
      example: 1
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: not found
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: connection refused
  schemas:
    Category:
      type: object
      required:
      - id
      - name
      - slug
      - product_count
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 1
          maxLength: 100
        slug:
          type: string
          minLength: 1
          maxLength: 100
        product_count:
          type: integer
          minimum: 0
    CategoryCreate:
      type: object
      required:
      - name
      - slug
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        slug:
          type: string
          minLength: 1
          maxLength: 100
    CategoryCreated:
      type: object
      required:
      - id
      - name
      - slug
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
        slug:
          type: string
    Product:
      type: object
      required:
      - id
      - name
      - price
      - stock
      - discount_pct
      - created_at
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
          minimum: 0
        stock:
          type: integer
          minimum: 0
        discount_pct:
          type: number
          format: float
          minimum: 0
          maximum: 100
        created_at:
          type: string
          format: date-time
        category_name:
          type: string
          nullable: true
        avg_rating:
          type: number
          format: float
          minimum: 1
          maximum: 5
          nullable: true
        review_count:
          type: integer
          minimum: 0
    ProductCreate:
      type: object
      required:
      - name
      - price
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          nullable: true
        category_id:
          type: integer
          minimum: 1
          nullable: true
        price:
          type: number
          format: float
          minimum: 0
        stock:
          type: integer
          minimum: 0
          default: 0
        discount_pct:
          type: number
          format: float
          minimum: 0
          maximum: 100
          default: 0
    ProductCreated:
      type: object
      required:
      - id
      - name
      - price
      - stock
      - discount_pct
      - created_at
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
        price:
          type: number
          format: float
        stock:
          type: integer
        discount_pct:
          type: number
          format: float
        created_at:
          type: string
          format: date-time
    ProductDetail:
      type: object
      required:
      - id
      - name
      - price
      - stock
      - discount_pct
      - created_at
      - reviews
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
        stock:
          type: integer
        discount_pct:
          type: number
          format: float
        created_at:
          type: string
          format: date-time
        category_name:
          type: string
          nullable: true
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
...
    Review:
      type: object
      required:
      - id
      - author
      - rating
      - created_at
      properties:
        id:
          type: integer
          minimum: 1
        author:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Stats:
      type: object
      required:
      - total_products
      - total_categories
      - total_reviews
      - total_orders
      - out_of_stock
      - low_stock
      - categories
      properties:
        total_products:
          type: integer
          minimum: 0
        total_categories:
          type: integer
          minimum: 0
        total_reviews:
          type: integer
          minimum: 0
        total_orders:
          type: integer
          minimum: 0
        out_of_stock:
          type: integer
          minimum: 0
        low_stock:
          type: integer
          minimum: 0
        avg_rating:
          type: number
          format: float
          minimum: 1
          maximum: 5
          nullable: true
        categories:
          type: array
          items:
            type: string
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          minLength: 1
...